backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
Pillow
opencv-python-headless
piexif
numpy
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pydantic-settings
python-dotenv
